payment complete

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,9 @@ use Illuminate\Support\Facades\Storage;
 use App\Models\User;
 use App\Models\Favorit;
 use App\Models\Item;
+use App\Models\Cart;
+use App\Models\Order;
+
 
 class UserController extends Controller
 {
@@ -28,12 +31,18 @@ class UserController extends Controller
 
     public function delete($email)
     {
-        $userCart = Cart::find($email);
-        $userCart->delete();
+        $userCart = Cart::where('user', $email);
+        if($userCart->count() > 0)
+        {
+            $userCart->delete();
+        }
 
         
-        $userFavorit = Favorit::find($email);
-        $userFavorit->delete();
+        $userFavorit = Favorit::where('user', $email);
+        if($userFavorit->count() > 0)
+        {
+            $userFavorit->delete();
+        }
 
         $user = User::find($email);
         Storage::delete($user->photo);
@@ -66,5 +75,38 @@ class UserController extends Controller
         $cekItem->save();
 
         return redirect()->route('profile.favorite');
+    }
+
+    public function status()
+    {
+        $title = 'FYP';
+        $active = 'Status';
+        $items = Order::all();
+
+        return view('status', compact('items', 'title', 'active'));
+    }
+
+    public function changeStatus(Request $request, $orderNum)
+    {
+        $orders = Order::where('orderNum', $orderNum)->get();
+        foreach ($orders as $order)
+        {
+            $order->status = $request->input('status');
+        }
+        $order->save();
+
+        return back();
+    }
+
+    public function deleteStatus($orderNum)
+    {
+        $order = Order::find($orderNum);
+        if(!$order->receipt == NULL)
+        {
+            Storage::delete($order->receipt);
+        }
+        $order->delete();
+
+        return back();
     }
 }
